requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
pycparser==2.21
pyspnego==0.9.0
PyYAML==6.0
requests==2.30.0
requests-ntlm==1.2.0
shellescape==3.8.1
tinydb==4.7.1
urllib3==2.0.2
